Added historical functionality to kafka.go
Updated PullFromTopic to take a sync channel and finalOffset as
additional parameters. The sync channel is used to signal the producers
for that partition that they should exit once they finish producing all
of their messages. The finalOffset is used to stop the consumer once it
has consumed all backlog data for the supplied fractional amount. If not
a historical run, it's a realtime run and finalOffset is -1 and therefore
the consumer never stops.